+++ e2e/file_es_split/docker-compose.yml
